itsgotime_api/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
